Rename parkinson filter and use relative file path
