--- src/Applied_Filtering_Methods/yes.py
@@ -1,1 +1,0 @@
-
